ibc: applications refactor
