+++ x/ibc/applications/transfer/keeper/genesis_test.go
@@ -3,7 +3,7 @@ package keeper_test
 import (
 	"fmt"
 
-	"github.com/cosmos/cosmos-sdk/x/ibc-transfer/types"
+	"github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/types"
 )
 
 func (suite *KeeperTestSuite) TestGenesis() {
