+++ x/ibc/applications/transfer/module.go
@@ -22,13 +22,13 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
-	"github.com/cosmos/cosmos-sdk/x/ibc-transfer/client/cli"
-	"github.com/cosmos/cosmos-sdk/x/ibc-transfer/keeper"
-	"github.com/cosmos/cosmos-sdk/x/ibc-transfer/simulation"
-	"github.com/cosmos/cosmos-sdk/x/ibc-transfer/types"
 	channeltypes "github.com/cosmos/cosmos-sdk/x/ibc/04-channel/types"
 	porttypes "github.com/cosmos/cosmos-sdk/x/ibc/05-port/types"
 	host "github.com/cosmos/cosmos-sdk/x/ibc/24-host"
+	"github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/client/cli"
+	"github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/keeper"
+	"github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/simulation"
+	"github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/types"
 )
 
 var (
