+++ x/ibc/applications/transfer/module_test.go
@@ -2,9 +2,9 @@ package transfer_test
 
 import (
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
-	"github.com/cosmos/cosmos-sdk/x/ibc-transfer/types"
 	channeltypes "github.com/cosmos/cosmos-sdk/x/ibc/04-channel/types"
 	host "github.com/cosmos/cosmos-sdk/x/ibc/24-host"
+	"github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/types"
 	ibctesting "github.com/cosmos/cosmos-sdk/x/ibc/testing"
 )
